week8/requirements.txt
matplotlib
numpy
pandas
QuantLib-Python
scikit-learn
torch
